IPC-101: Move Client to its own module.
