+++ ipld/resolver/src/lib.rs
@@ -1,6 +1,7 @@
 // Copyright 2022-2023 Protocol Labs
 // SPDX-License-Identifier: MIT
 mod behaviour;
+mod client;
 mod hash;
 mod limiter;
 mod provider_cache;
@@ -15,4 +16,5 @@ mod arb;
 pub mod missing_blocks;
 
 pub use behaviour::{ContentConfig, DiscoveryConfig, MembershipConfig, NetworkConfig};
-pub use service::{Client, Config, ConnectionConfig, NoKnownPeers, Service};
+pub use client::Client;
+pub use service::{Config, ConnectionConfig, NoKnownPeers, Service};
